+++ src/gort/recipes/operations.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 
 import asyncio
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, ClassVar
 
 from rich.prompt import Confirm
 
@@ -322,7 +322,9 @@ class PostObservingRecipe(BaseRecipe):
 
     name = "post-observing"
 
-    async def recipe(self):
+    email_route: ClassVar[str] = "/logs/night-logs/0/email?only_if_not_sent=1"
+
+    async def recipe(self, send_email: bool = True):
         """Runs the post-observing sequence."""
 
         from gort.overwatcher.helpers.notifier import BasicNotifier
@@ -347,10 +349,11 @@ class PostObservingRecipe(BaseRecipe):
             except Exception as ee:
                 notifier.log.error(f"Error running post-observing task: {ee}")
 
-        notifier.log.info("Sending night log email.")
-        result = await get_lvmapi_route("/logs/night-logs/0/email?only_if_not_sent=1")
-        if not result:
-            notifier.log.warning("Night log had already been sent.")
+        if send_email:
+            notifier.log.info("Sending night log email.")
+            result = await get_lvmapi_route(self.email_route)
+            if not result:
+                notifier.log.warning("Night log had already been sent.")
 
         # Disable the overwatcher.
         if await overwatcher_is_running():